Refactors to allow multiple LLM providers
Refactors the code to move the LLM generatin logic out of the route
definition and into a separate module. This allows for easier
extensibility and the addition of new LLM providers in the future.

This commit also adds a new Gemini provider, which is a wrapper
around the Gemini API.